task 4
Using an MD5 algorithm to hash the input files, then they get cracked through collision free and one way methods. The cracks are ran and compared to the hashes of the original inputs until matches are made. 

Final step was to:
- find two lower case ascii inputs which hash to the same value
- given a 24-bit hash value stored as a 6 character hex hash (for matching input files through cracking)
- so the program will have to repeatedly generate a random text, hash it, then compare
lower 24 bits to the original input of your program